Revert "adding to_do list file"
