+++ lib/main.dart
@@ -1,14 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:zen/auth_pages/auth_page.dart';
+import 'package:zen/pages/auth_page.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:zen/screens/home.dart';
-import 'package:zen/screens/mood.dart';
-import 'package:zen/screens/todo.dart';
 import 'firebase_options.dart';
+
 
 void main() async {
   //to ensure firebase plugins are correctly intialised before using it
-  WidgetsFlutterBinding.ensureInitialized();
+  WidgetsFlutterBinding.ensureInitialized(); 
   await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
   runApp(const MyApp());
 }
@@ -18,15 +16,11 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       debugShowCheckedModeBanner: false,
-      initialRoute: '/', //Specifies the initial page route
-      routes: {
-        '/': (context) => const AuthPage(),
-        '/home': (context) => const LandPage(),
-        '/mood': (context) => const MoodPage(),
-        '/todo': (context) => const TodoPage(),
-      },
+      home: AuthPage(),
     );
   }
 }
+
+
