+++ lib/pages/auth_page.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:zen/auth_pages/login_page.dart';
-import 'package:zen/screens/home.dart';
+import 'package:zen/pages/home_page.dart';
+import 'package:zen/pages/login_page.dart';
 
 class AuthPage extends StatelessWidget {
   const AuthPage({super.key});
@@ -12,10 +12,11 @@ class AuthPage extends StatelessWidget {
         body: StreamBuilder(
             stream: FirebaseAuth.instance.authStateChanges(),
             builder: (context, snapshot) {
-              //if user is logged in
+
+              //if user is logged in 
               if (snapshot.hasData) {
-                return const LandPage();
-              }
+                return const HomePage();
+              } 
               //if user is not logged in
               else {
                 return const LoginPage();
