--- lib/screens/mood.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MoodPage extends StatelessWidget {
-  const MoodPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Text('Mood Page'),
-      ),
-    );
-  }
-}
