--- lib/screens/todo.dart
@@ -1,39 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:zen/pages/home_page.dart';
-
-class Todo extends StatefulWidget {
-  const Todo({super.key});
-
-  @override
-  State<Todo> createState() => _TodoState();
-}
-
-class _TodoState extends State<Todo> {
-  final TextEditingController nameController = TextEditingController();
-  final TextEditingController descController = TextEditingController();
-  final TextEditingController dateController = TextEditingController();
-  final TextEditingController priorityController = TextEditingController();
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("To-Do list",
-            style: TextStyle(
-                color: Colors.black,
-                fontSize: 26,
-                fontWeight: FontWeight.bold)),
-        backgroundColor: Colors.blue,
-        actions: [
-          IconButton(
-              onPressed: () {
-                Navigator.pushReplacement(context,
-                    MaterialPageRoute(builder: (context) => HomePage()));
-              },
-              icon: Icon(Icons.arrow_back)),
-        ],
-      ),
-      //to add details....
-      //body: ,
-    );
-  }
-}
